observability/requirements.txt
# Observability Service Requirements
fastapi
uvicorn[standard]
pydantic
pydantic-settings
sqlalchemy
psycopg2-binary
alembic
redis
aioredis
prometheus-client
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-sqlalchemy
opentelemetry-instrumentation-psycopg2
opentelemetry-instrumentation-redis
opentelemetry-exporter-jaeger
opentelemetry-exporter-prometheus
deprecated
structlog
python-json-logger
aiohttp
httpx
python-multipart
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
tenacity
asyncpg
psutil
websockets

